worker/requirements.txt
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.3
google-cloud-firestore==2.11.1
google-cloud-pubsub==2.18.4
GitPython==3.1.40
PyGithub==2.1.1
Jinja2==3.1.2
python-dotenv==1.0.0

